[Rename] ui-actions plugin
