--- src/plugins/ui_actions/kibana.json
@@ -1,10 +1,0 @@
-{
-  "id": "uiActions",
-  "version": "kibana",
-  "server": false,
-  "ui": true,
-  "requiredBundles": [
-    "kibanaUtils",
-    "kibanaReact"
-  ]
-}
