+++ src/plugins/ui_actions/public/actions/incompatible_action_error.ts
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { i18n } from '@kbn/i18n';
+import { i18n } from '@osd/i18n';
 
 export class IncompatibleActionError extends Error {
   code = 'INCOMPATIBLE_ACTION';
