+++ src/plugins/ui_actions/public/context_menu/build_eui_context_menu_panels.tsx
@@ -20,8 +20,8 @@
 import * as React from 'react';
 import { EuiContextMenuPanelDescriptor, EuiContextMenuPanelItemDescriptor } from '@elastic/eui';
 import _ from 'lodash';
-import { i18n } from '@kbn/i18n';
-import { uiToReactComponent } from '../../../kibana_react/public';
+import { i18n } from '@osd/i18n';
+import { uiToReactComponent } from '../../../opensearch_dashboards_react/public';
 import { Action, ActionExecutionContext } from '../actions';
 import { Trigger } from '../triggers';
 import { BaseContext } from '../types';
