+++ src/plugins/ui_actions/public/service/ui_actions_execution_service.ts
@@ -20,7 +20,7 @@
 import { uniqBy } from 'lodash';
 import { Action } from '../actions';
 import { BaseContext } from '../types';
-import { defer as createDefer, Defer } from '../../../kibana_utils/public';
+import { defer as createDefer, Defer } from '../../../opensearch_dashboards_utils/public';
 import { buildContextMenuForActions, openContextMenu } from '../context_menu';
 import { Trigger } from '../triggers';
