+++ src/plugins/ui_actions/public/tests/test_samples/hello_world_action.tsx
@@ -21,7 +21,7 @@ import React from 'react';
 import { EuiFlexGroup, EuiFlexItem, EuiBadge, EuiFlyoutBody } from '@elastic/eui';
 import { CoreStart } from 'src/core/public';
 import { createAction, ActionByType } from '../../actions';
-import { toMountPoint, reactToUiComponent } from '../../../../kibana_react/public';
+import { toMountPoint, reactToUiComponent } from '../../../../opensearch_dashboards_react/public';
 import { ActionType } from '../../types';
 
 const ReactMenuItem: React.FC = () => {
