+++ src/plugins/ui_actions/public/triggers/select_range_trigger.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { i18n } from '@kbn/i18n';
+import { i18n } from '@osd/i18n';
 import { Trigger } from '.';
 
 export const SELECT_RANGE_TRIGGER = 'SELECT_RANGE_TRIGGER';
